# swagger.yaml
openapi: 3.0.3
info:
  title: Blue Nest Montessori API
  description: |
    REST API for the Blue Nest Montessori nursery store and blog platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    `Authorization: Bearer <access_token>`
    
    ## Roles
    - **customer**: Regular authenticated users
    - **admin**: Administrative users with full access
    - **branch_manager**: Branch-specific management access
  version: 1.0.0
  contact:
    name: Blue Nest Montessori
    url: https://blue-nest-montessori.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.blue-nest-montessori.com/api/v1
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register new user
      description: Create a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - first_name
                - last_name
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: securePassword123
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request body
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user and return JWT tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      tags:
        - Auth
      summary: User logout
      description: Invalidate the current session
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Use refresh token to obtain new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '401':
          description: Invalid or expired refresh token

  # ─────────────────────────────────────────────────────────────────────────────
  # Products
  # ─────────────────────────────────────────────────────────────────────────────
  /products:
    get:
      tags:
        - Products
      summary: List all products
      description: Returns a paginated list of active products
      operationId: listProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: branch
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                  page:
                    type: integer
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Returns a single product by its ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
...
  /blog/posts:
    get:
      tags:
        - Blog
      summary: List blog posts
      description: Returns a list of published blog posts
      operationId: listBlogPosts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: tag
          in: query
          schema:
            type: string
        - name: branch
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of blog posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlogPost'
                  total:
                    type: integer
                  page:
                    type: integer
  /blog/posts/{slug}:
    get:
      tags:
        - Blog
      summary: Get blog post by slug
      description: Returns a single blog post by its URL slug
      operationId: getBlogPost
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            example: my-first-post
      responses:
        '200':
          description: Blog post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          description: Blog post not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Branches
  # ─────────────────────────────────────────────────────────────────────────────
...
  /branches/{slug}:
    get:
      tags:
        - Branches
      summary: Get branch by slug
      description: Returns a single branch by its URL slug
      operationId: getBranch
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            example: borehamwood
      responses:
        '200':
          description: Branch details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        '404':
          description: Branch not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Cart (Authenticated)
  # ─────────────────────────────────────────────────────────────────────────────
  /cart:
    get:
      tags:
        - Cart
      summary: Get user cart
      description: Returns the current user's shopping cart
      operationId: getCart
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cart details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /cart/items:
    post:
      tags:
        - Cart
      summary: Add item to cart
      description: Add a product to the shopping cart
      operationId: addCartItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - qty
              properties:
                product_id:
                  type: string
                  example: 507f1f77bcf86cd799439011
                qty:
                  type: integer
                  minimum: 1
                  example: 2
      responses:
        '201':
          description: Item added to cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid request
        '404':
          description: Product not found
  /cart/items/{id}:
    put:
      tags:
        - Cart
      summary: Update cart item quantity
      description: Update the quantity of an item in the cart
      operationId: updateCartItem
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - qty
              properties:
                qty:
                  type: integer
                  minimum: 0
                  example: 3
      responses:
        '200':
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Cart item not found

    delete:
      tags:
        - Cart
      summary: Remove cart item
      description: Remove an item from the shopping cart
      operationId: removeCartItem
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Item removed from cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Cart item not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Checkout (Authenticated)
  # ─────────────────────────────────────────────────────────────────────────────
  /checkout/session:
    post:
      tags:
        - Checkout
      summary: Create checkout session
      description: Create a Stripe checkout session for the current cart
      operationId: createCheckoutSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - success_url
                - cancel_url
              properties:
                success_url:
                  type: string
                  format: uri
                  example: https://blue-nest-montessori.com/checkout/success
                cancel_url:
                  type: string
                  format: uri
                  example: https://blue-nest-montessori.com/checkout/cancel
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  url:
                    type: string
                    format: uri
        '400':
          description: Cart is empty or invalid

  # ─────────────────────────────────────────────────────────────────────────────
  # Orders (Authenticated)
  # ─────────────────────────────────────────────────────────────────────────────
  /orders/me:
    get:
      tags:
        - Orders
      summary: List user's orders
      description: Returns all orders for the authenticated user
      operationId: listMyOrders
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of user's orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Returns a single order by its ID
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Not authorized to view this order
        '404':
          description: Order not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Admin - Orders
  # ─────────────────────────────────────────────────────────────────────────────
  /admin/orders:
    get:
      tags:
        - Admin
      summary: List all orders (admin)
      description: Returns all orders across all users
      operationId: adminListOrders
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, processing, shipped, delivered, cancelled]
      responses:
        '200':
          description: List of all orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
  /admin/orders/{id}:
    get:
      tags:
        - Admin
      summary: Get order by ID (admin)
      description: Returns a single order by its ID
      operationId: adminGetOrder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
  /admin/orders/{id}/status:
    patch:
      tags:
        - Admin
      summary: Update order status
      description: Update the status of an order
      operationId: updateOrderStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, paid, processing, shipped, delivered, cancelled]
                  example: processing
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status
        '404':
          description: Order not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Admin - Products
  # ─────────────────────────────────────────────────────────────────────────────
  /admin/products:
    get:
      tags:
        - Admin
      summary: List all products (admin)
      description: Returns all products including inactive ones
      operationId: adminListProducts
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer

    post:
      tags:
        - Admin
      summary: Create product
      description: Create a new product
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /admin/products/{id}:
    put:
      tags:
        - Admin
      summary: Update product
      description: Update an existing product
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

    delete:
      tags:
        - Admin
      summary: Delete product
      description: Delete a product
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '204':
          description: Product deleted

  # ─────────────────────────────────────────────────────────────────────────────
  # Admin - Blog
  # ─────────────────────────────────────────────────────────────────────────────
  /admin/blog/posts:
    get:
      tags:
        - Admin
      summary: List all blog posts (admin)
      description: Returns all blog posts including drafts
      operationId: adminListBlogPosts
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of all blog posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlogPost'
                  total:
                    type: integer

    post:
      tags:
        - Admin
      summary: Create blog post
      description: Create a new blog post
      operationId: createBlogPost
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPostInput'
      responses:
        '201':
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
  /admin/blog/posts/{id}:
    put:
      tags:
        - Admin
      summary: Update blog post
      description: Update an existing blog post
      operationId: updateBlogPost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPostInput'
      responses:
        '200':
          description: Blog post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          description: Blog post not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Admin - Users (super-admin account management)
  # ─────────────────────────────────────────────────────────────────────────────
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users (super-admin)
      operationId: adminListUsers
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of users }
        '403': { description: Super-admin only }
    post:
      tags: [Admin]
      summary: Create a user (super-admin)
      operationId: adminCreateUser
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, first_name, last_name, role]
              properties:
                email: { type: string }
                password: { type: string, minLength: 8 }
                first_name: { type: string }
                last_name: { type: string }
                role:
                  type: string
                  enum: [super_admin, admin, branch_manager, staff, customer]
      responses:
        '201': { description: User created }
        '400': { description: Invalid input }
  /admin/users/{id}:
    put:
      tags: [Admin]
      summary: Update a user (super-admin). Optional password field also resets the password.
      operationId: adminUpdateUser
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: { type: string }
                last_name: { type: string }
                role:
                  type: string
                  enum: [super_admin, admin, branch_manager, staff, customer]
                password: { type: string, minLength: 8 }
      responses:
        '200': { description: User updated }
        '400': { description: Invalid input or self-demotion blocked }
    delete:
      tags: [Admin]
      summary: Delete a user (super-admin). Cannot delete your own account.
      operationId: adminDeleteUser
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: User deleted }
        '400': { description: Self-delete blocked }
...
  /order-requests/me:
    get:
      tags: [Order Requests]
      summary: List my supply requests
      operationId: listMyOrderRequests
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of the caller's requests }
...
  /admin/order-requests:
    get:
      tags: [Admin]
      summary: List all supply requests (management)
      operationId: adminListOrderRequests
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of all requests }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'

    # ─────────────────────────────────────────────────────────────────────────
    # Products
    # ─────────────────────────────────────────────────────────────────────────
...